+++ src/components/Nodes/nodeHelpers.ts
@@ -1,0 +1,21 @@
+type Node = {
+  status: {
+    allocatable: any;
+  };
+};
+
+export function getAvailableNvidiaGPUs(nodesData: Node[]): Number {
+  if (!nodesData) {
+    return 0;
+  }
+  return nodesData
+    .map(node => node.status.allocatable)
+    .reduce((partialSum, item) => {
+      const nvidiaGpus = item['nvidia.com/gpu'];
+      if (nvidiaGpus) {
+        const value = parseInt(nvidiaGpus);
+        return partialSum + value;
+      }
+      return partialSum;
+    }, 0);
+}
