+++ src/resources/Namespaces/NamespaceWorkloads/NamespaceWorkloadsHelpers.js
@@ -63,19 +63,3 @@ export function getHealthyDaemonsets(daemonsets) {
       ds.status.numberMisscheduled === 0,
   )?.length;
 }
-
-export function getAvailableNvidiaGPUs(nodesData) {
-  if (!nodesData) {
-    return 0;
-  }
-  return nodesData
-    .map(node => node.status.allocatable)
-    .reduce((partialSum, item) => {
-      const nvidiaGpus = item['nvidia.com/gpu'];
-      if (nvidiaGpus) {
-        const value = parseInt(nvidiaGpus);
-        return partialSum + value;
-      }
-      return partialSum;
-    }, 0);
-}
